Create elfFusion.c and update doc
