+++ doc/html/elfFile_8h.html
@@ -98,12 +98,16 @@ Functions</h2></td></tr>
 <tr class="separator:acdb032fddcba7d654a5d0d3456a3a7be"><td class="memSeparator" colspan="2">&#160;</td></tr>
 <tr class="memitem:a41b315ed88b2766ba8174206562c419d"><td class="memItemLeft" align="right" valign="top">void&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#a41b315ed88b2766ba8174206562c419d">ShowReimplantationTablesAndDetails</a> (FILE *elfFile, <a class="el" href="structElf32__Structure.html">Elf32_Structure</a> structure)</td></tr>
 <tr class="separator:a41b315ed88b2766ba8174206562c419d"><td class="memSeparator" colspan="2">&#160;</td></tr>
+<tr class="memitem:a941f7b8cd416ad6c103dd50a60e3152e"><td class="memItemLeft" align="right" valign="top">void&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#a941f7b8cd416ad6c103dd50a60e3152e">DoFusionCommand</a> (FILE **elfFiles, <a class="el" href="structElf32__Structure.html">Elf32_Structure</a> *structure)</td></tr>
+<tr class="separator:a941f7b8cd416ad6c103dd50a60e3152e"><td class="memSeparator" colspan="2">&#160;</td></tr>
 <tr class="memitem:ab5003ae56f77c5d7fdf49fe349d307a2"><td class="memItemLeft" align="right" valign="top"><a class="el" href="structElf32__SectionFusion.html">Elf32_SectionFusion</a>&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#ab5003ae56f77c5d7fdf49fe349d307a2">FusionSections</a> (FILE **elfFiles, Elf32_Ehdr *elfHeaders, <a class="el" href="elfStructure_8h.html#af909f76c31b2301c5623e615956db471">Elf32_ShdrTable</a> *sectionTables)</td></tr>
 <tr class="separator:ab5003ae56f77c5d7fdf49fe349d307a2"><td class="memSeparator" colspan="2">&#160;</td></tr>
 <tr class="memitem:a97d7c187490aea61028153a39adfab6b"><td class="memItemLeft" align="right" valign="top"><a class="el" href="structElf32__SymbolFusion.html">Elf32_SymbolFusion</a>&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#a97d7c187490aea61028153a39adfab6b">FusionSymbols</a> (FILE *elfFiles[2], Elf32_Ehdr elfHeaders[2], <a class="el" href="elfStructure_8h.html#ad9a7d41d2b4a95d426c75f16b926e0f7">Elf32_SymTable</a> symbolTables[2], <a class="el" href="structElf32__SectionFusion.html">Elf32_SectionFusion</a> sectionFusion)</td></tr>
 <tr class="separator:a97d7c187490aea61028153a39adfab6b"><td class="memSeparator" colspan="2">&#160;</td></tr>
 <tr class="memitem:a3144f86bfcb8a1f1b415ba4df07f8688"><td class="memItemLeft" align="right" valign="top"><a class="el" href="structElf32__RelFusion.html">Elf32_RelFusion</a>&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#a3144f86bfcb8a1f1b415ba4df07f8688">FusionReimplantation</a> (FILE **elfFiles, <a class="el" href="structElf32__Structure.html">Elf32_Structure</a> *structure, <a class="el" href="structElf32__SectionFusion.html">Elf32_SectionFusion</a> sectionFusion, <a class="el" href="structElf32__SymbolFusion.html">Elf32_SymbolFusion</a> symbolFusion)</td></tr>
 <tr class="separator:a3144f86bfcb8a1f1b415ba4df07f8688"><td class="memSeparator" colspan="2">&#160;</td></tr>
+<tr class="memitem:a97ee70a8770dc30d06c744b24eb2fcfc"><td class="memItemLeft" align="right" valign="top">void&#160;</td><td class="memItemRight" valign="bottom"><a class="el" href="elfFile_8h.html#a97ee70a8770dc30d06c744b24eb2fcfc">help</a> ()</td></tr>
+<tr class="separator:a97ee70a8770dc30d06c744b24eb2fcfc"><td class="memSeparator" colspan="2">&#160;</td></tr>
 </table><table class="memberdecls">
 <tr class="heading"><td colspan="2"><h2 class="groupheader"><a id="var-members" name="var-members"></a>
 Variables</h2></td></tr>
@@ -113,6 +117,48 @@ Variables</h2></td></tr>
 <a name="details" id="details"></a><h2 class="groupheader">Detailed Description</h2>
 <div class="textblock"><p >Fichier d'en-tête principal. Contient les signatures des méthodes d'extractions et d'affichage </p>
 </div><h2 class="groupheader">Function Documentation</h2>
+<a id="a941f7b8cd416ad6c103dd50a60e3152e" name="a941f7b8cd416ad6c103dd50a60e3152e"></a>
+<h2 class="memtitle"><span class="permalink"><a href="#a941f7b8cd416ad6c103dd50a60e3152e">&#9670;&nbsp;</a></span>DoFusionCommand()</h2>
+
+<div class="memitem">
+<div class="memproto">
+      <table class="memname">
+        <tr>
+          <td class="memname">void DoFusionCommand </td>
+          <td>(</td>
+          <td class="paramtype">FILE **&#160;</td>
+          <td class="paramname"><em>elfFiles</em>, </td>
+        </tr>
+        <tr>
+          <td class="paramkey"></td>
+          <td></td>
+          <td class="paramtype"><a class="el" href="structElf32__Structure.html">Elf32_Structure</a> *&#160;</td>
+          <td class="paramname"><em>structure</em>&#160;</td>
+        </tr>
+        <tr>
+          <td></td>
+          <td>)</td>
+          <td></td><td></td>
+        </tr>
+      </table>
+</div><div class="memdoc">
+<p >Fusionne les 2 fichiers elf <br  />
+ Précondition : <br  />
+</p><ul>
+<li><code>elfFiles</code> doit être un tableau de 2 fichiers ouverts en lecture <br  />
+</li>
+<li><code>structure</code> doit être un tableau de 2 structures <dl class="params"><dt>Parameters</dt><dd>
+  <table class="params">
+    <tr><td class="paramname">elfFiles</td><td>Fichiers elf à fusionner </td></tr>
+    <tr><td class="paramname">structure</td><td>Structure des 2 fichiers elf </td></tr>
+  </table>
+  </dd>
+</dl>
+</li>
+</ul>
+
+</div>
+</div>
 <a id="a19ae190d378c6090b327ca358c130861" name="a19ae190d378c6090b327ca358c130861"></a>
 <h2 class="memtitle"><span class="permalink"><a href="#a19ae190d378c6090b327ca358c130861">&#9670;&nbsp;</a></span>ExtractElfInformation()</h2>
 
@@ -464,6 +510,24 @@ Variables</h2></td></tr>
   </dd>
 </dl>
 <dl class="section return"><dt>Returns</dt><dd>Le résultat de la fusion </dd></dl>
+
+</div>
+</div>
+<a id="a97ee70a8770dc30d06c744b24eb2fcfc" name="a97ee70a8770dc30d06c744b24eb2fcfc"></a>
+<h2 class="memtitle"><span class="permalink"><a href="#a97ee70a8770dc30d06c744b24eb2fcfc">&#9670;&nbsp;</a></span>help()</h2>
+
+<div class="memitem">
+<div class="memproto">
+      <table class="memname">
+        <tr>
+          <td class="memname">void help </td>
+          <td>(</td>
+          <td class="paramname"></td><td>)</td>
+          <td></td>
+        </tr>
+      </table>
+</div><div class="memdoc">
+<p >Affiche l'aide du programme </p>
 
 </div>
 </div>
