+++ doc/html/elfStructure_8h_source.html
@@ -150,7 +150,7 @@ $(function() {
 <div class="line"><a id="l00146" name="l00146"></a><span class="lineno">  146</span> </div>
 <div class="line"><a id="l00147" name="l00147"></a><span class="lineno">  147</span><span class="preprocessor">#endif </span><span class="comment">//PROJET_PROG_ELFSTRUCTURE_H</span></div>
 <div class="ttc" id="aelfStructure_8h_html_a469032e68c22b28f7a91100540b6c9ae"><div class="ttname"><a href="elfStructure_8h.html#a469032e68c22b28f7a91100540b6c9ae">NewSectionFusion</a></div><div class="ttdeci">Elf32_SectionFusion NewSectionFusion(Elf32_Word sectionSize1, Elf32_Word sectionSize2)</div><div class="ttdef"><b>Definition:</b> elfStructure.c:10</div></div>
-<div class="ttc" id="aelfStructure_8h_html_a5a5fb7b2384edf089e3e7c61393e14be"><div class="ttname"><a href="elfStructure_8h.html#a5a5fb7b2384edf089e3e7c61393e14be">FreeElf32_Structure</a></div><div class="ttdeci">void FreeElf32_Structure(Elf32_Structure structure)</div><div class="ttdef"><b>Definition:</b> elfStructure.c:44</div></div>
+<div class="ttc" id="aelfStructure_8h_html_a5a5fb7b2384edf089e3e7c61393e14be"><div class="ttname"><a href="elfStructure_8h.html#a5a5fb7b2384edf089e3e7c61393e14be">FreeElf32_Structure</a></div><div class="ttdeci">void FreeElf32_Structure(Elf32_Structure structure)</div><div class="ttdef"><b>Definition:</b> elfStructure.c:48</div></div>
 <div class="ttc" id="aelfStructure_8h_html_a5e81e8e9f72fd8613ea01fbf2b9e1e8e"><div class="ttname"><a href="elfStructure_8h.html#a5e81e8e9f72fd8613ea01fbf2b9e1e8e">NewElf32_Structure</a></div><div class="ttdeci">Elf32_Structure NewElf32_Structure(char *path, Elf32_Ehdr header, Elf32_ShdrTable sectionTable, Elf32_SymTable symbolTable, Elf32_ReimTable reimplantationTable, int sectionCount, int symbolCount, int reimplantationCount)</div><div class="ttdef"><b>Definition:</b> elfStructure.c:35</div></div>
 <div class="ttc" id="aelfStructure_8h_html_a747b86ad1ca395db95235b7d51c0c88e"><div class="ttname"><a href="elfStructure_8h.html#a747b86ad1ca395db95235b7d51c0c88e">FreeSectionFusion</a></div><div class="ttdeci">void FreeSectionFusion(Elf32_SectionFusion fusion)</div><div class="ttdef"><b>Definition:</b> elfStructure.c:28</div></div>
 <div class="ttc" id="aelfStructure_8h_html_ad9a7d41d2b4a95d426c75f16b926e0f7"><div class="ttname"><a href="elfStructure_8h.html#ad9a7d41d2b4a95d426c75f16b926e0f7">Elf32_SymTable</a></div><div class="ttdeci">Elf32_Sym * Elf32_SymTable</div><div class="ttdef"><b>Definition:</b> elfStructure.h:21</div></div>
