+++ doc/html/globals.html
@@ -62,8 +62,19 @@ $(function() {
 </div>
 
 <div class="contents">
-<div class="textblock">Here is a list of all documented functions, variables, defines, enums, and typedefs with links to the documentation:</div><ul>
+<div class="textblock">Here is a list of all documented functions, variables, defines, enums, and typedefs with links to the documentation:</div>
+
+<h3><a id="index_c" name="index_c"></a>- c -</h3><ul>
 <li>CheckMachineEndian()&#160;:&#160;<a class="el" href="utils_8h.html#a88bf933dd8e43ab559d8539c7c67b3b6">utils.h</a></li>
+</ul>
+
+
+<h3><a id="index_d" name="index_d"></a>- d -</h3><ul>
+<li>DoFusionCommand()&#160;:&#160;<a class="el" href="elfFile_8h.html#a941f7b8cd416ad6c103dd50a60e3152e">elfFile.h</a></li>
+</ul>
+
+
+<h3><a id="index_e" name="index_e"></a>- e -</h3><ul>
 <li>Elf32_ReimTable&#160;:&#160;<a class="el" href="elfStructure_8h.html#ae2ea0ee910877c478579d0e44777d18d">elfStructure.h</a></li>
 <li>Elf32_RelaTable&#160;:&#160;<a class="el" href="elfStructure_8h.html#aea239d6afe7de034af35bc834cc22dfa">elfStructure.h</a></li>
 <li>Elf32_ShdrTable&#160;:&#160;<a class="el" href="elfStructure_8h.html#af909f76c31b2301c5623e615956db471">elfStructure.h</a></li>
@@ -73,19 +84,44 @@ $(function() {
 <li>ExtractReimplantationTable()&#160;:&#160;<a class="el" href="elfFile_8h.html#a56244c69ae5f37064f10f780b60cb1b7">elfFile.h</a></li>
 <li>ExtractSectionTable()&#160;:&#160;<a class="el" href="elfFile_8h.html#ac854b60ad3231b4b2618433475e96ee8">elfFile.h</a></li>
 <li>ExtractSymbolsTable()&#160;:&#160;<a class="el" href="elfFile_8h.html#a1fffc3393c1db5b07e408be35601b9c0">elfFile.h</a></li>
+</ul>
+
+
+<h3><a id="index_f" name="index_f"></a>- f -</h3><ul>
 <li>freadEndian()&#160;:&#160;<a class="el" href="utils_8h.html#a0479d547af675b99a4dd3df59d5d3604">utils.h</a></li>
 <li>FreeElf32_Structure()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a5a5fb7b2384edf089e3e7c61393e14be">elfStructure.h</a></li>
 <li>FreeSectionFusion()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a747b86ad1ca395db95235b7d51c0c88e">elfStructure.h</a></li>
 <li>FusionReimplantation()&#160;:&#160;<a class="el" href="elfFile_8h.html#a3144f86bfcb8a1f1b415ba4df07f8688">elfFile.h</a></li>
 <li>FusionSections()&#160;:&#160;<a class="el" href="elfFile_8h.html#ab5003ae56f77c5d7fdf49fe349d307a2">elfFile.h</a></li>
 <li>FusionSymbols()&#160;:&#160;<a class="el" href="elfFile_8h.html#a97d7c187490aea61028153a39adfab6b">elfFile.h</a></li>
+</ul>
+
+
+<h3><a id="index_g" name="index_g"></a>- g -</h3><ul>
 <li>GetEntryCountFromType()&#160;:&#160;<a class="el" href="utils_8h.html#ae960326a7cdebf3af94ee2ff87076d7c">utils.h</a></li>
 <li>GetSectionCountFromType()&#160;:&#160;<a class="el" href="utils_8h.html#a569d0f2da13d385132706f3f506c37fa">utils.h</a></li>
 <li>GetSectionIndexByName()&#160;:&#160;<a class="el" href="utils_8h.html#a2f0ff914ecaef8d1f120ece67cd8678a">utils.h</a></li>
+</ul>
+
+
+<h3><a id="index_h" name="index_h"></a>- h -</h3><ul>
+<li>help()&#160;:&#160;<a class="el" href="elfFile_8h.html#a97ee70a8770dc30d06c744b24eb2fcfc">elfFile.h</a></li>
+</ul>
+
+
+<h3><a id="index_m" name="index_m"></a>- m -</h3><ul>
 <li>mallocArray&#160;:&#160;<a class="el" href="utils_8h.html#a5a6fb16fa1bb3f3085c48b7106a5be59">utils.h</a></li>
+</ul>
+
+
+<h3><a id="index_n" name="index_n"></a>- n -</h3><ul>
 <li>needReverse&#160;:&#160;<a class="el" href="elfFile_8h.html#ab377c00452ebed00f3024890d27ddb7e">elfFile.h</a></li>
 <li>NewElf32_Structure()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a5e81e8e9f72fd8613ea01fbf2b9e1e8e">elfStructure.h</a></li>
 <li>NewSectionFusion()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a469032e68c22b28f7a91100540b6c9ae">elfStructure.h</a></li>
+</ul>
+
+
+<h3><a id="index_s" name="index_s"></a>- s -</h3><ul>
 <li>ShowElfHeader()&#160;:&#160;<a class="el" href="elfFile_8h.html#a913a20ad3ffeff20afe960474ec7d34b">elfFile.h</a></li>
 <li>ShowReimplantationTablesAndDetails()&#160;:&#160;<a class="el" href="elfFile_8h.html#a41b315ed88b2766ba8174206562c419d">elfFile.h</a></li>
 <li>ShowSectionFromIndex()&#160;:&#160;<a class="el" href="elfFile_8h.html#a71973292791ceabf3f391980c1478370">elfFile.h</a></li>
