+++ doc/html/globals_func.html
@@ -64,6 +64,7 @@ $(function() {
 <div class="contents">
 &#160;<ul>
 <li>CheckMachineEndian()&#160;:&#160;<a class="el" href="utils_8h.html#a88bf933dd8e43ab559d8539c7c67b3b6">utils.h</a></li>
+<li>DoFusionCommand()&#160;:&#160;<a class="el" href="elfFile_8h.html#a941f7b8cd416ad6c103dd50a60e3152e">elfFile.h</a></li>
 <li>ExtractElfInformation()&#160;:&#160;<a class="el" href="elfFile_8h.html#a19ae190d378c6090b327ca358c130861">elfFile.h</a></li>
 <li>ExtractHeader()&#160;:&#160;<a class="el" href="elfFile_8h.html#a27ba046fdd1470f335e7379cbcab8f50">elfFile.h</a></li>
 <li>ExtractReimplantationTable()&#160;:&#160;<a class="el" href="elfFile_8h.html#a56244c69ae5f37064f10f780b60cb1b7">elfFile.h</a></li>
@@ -78,6 +79,7 @@ $(function() {
 <li>GetEntryCountFromType()&#160;:&#160;<a class="el" href="utils_8h.html#ae960326a7cdebf3af94ee2ff87076d7c">utils.h</a></li>
 <li>GetSectionCountFromType()&#160;:&#160;<a class="el" href="utils_8h.html#a569d0f2da13d385132706f3f506c37fa">utils.h</a></li>
 <li>GetSectionIndexByName()&#160;:&#160;<a class="el" href="utils_8h.html#a2f0ff914ecaef8d1f120ece67cd8678a">utils.h</a></li>
+<li>help()&#160;:&#160;<a class="el" href="elfFile_8h.html#a97ee70a8770dc30d06c744b24eb2fcfc">elfFile.h</a></li>
 <li>NewElf32_Structure()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a5e81e8e9f72fd8613ea01fbf2b9e1e8e">elfStructure.h</a></li>
 <li>NewSectionFusion()&#160;:&#160;<a class="el" href="elfStructure_8h.html#a469032e68c22b28f7a91100540b6c9ae">elfStructure.h</a></li>
 <li>ShowElfHeader()&#160;:&#160;<a class="el" href="elfFile_8h.html#a913a20ad3ffeff20afe960474ec7d34b">elfFile.h</a></li>
