+++ doc/html/search/all_4.js
@@ -1,4 +1,6 @@
 var searchData=
 [
-  ['header_0',['header',['../structElf32__Structure.html#a8f94f97699ff16c74a742bebd5b4da58',1,'Elf32_Structure']]]
+  ['getentrycountfromtype_0',['GetEntryCountFromType',['../utils_8h.html#ae960326a7cdebf3af94ee2ff87076d7c',1,'utils.c']]],
+  ['getsectioncountfromtype_1',['GetSectionCountFromType',['../utils_8h.html#a569d0f2da13d385132706f3f506c37fa',1,'utils.c']]],
+  ['getsectionindexbyname_2',['GetSectionIndexByName',['../utils_8h.html#a2f0ff914ecaef8d1f120ece67cd8678a',1,'utils.c']]]
 ];
