+++ doc/html/search/all_8.js
@@ -1,4 +1,9 @@
 var searchData=
 [
-  ['path_0',['path',['../structElf32__Structure.html#a44196e6a5696d10442c29e639437196e',1,'Elf32_Structure']]]
+  ['nbrel_0',['nbRel',['../structElf32__Reim.html#ae8b0bd33116a39397523c6a0a828adaa',1,'Elf32_Reim']]],
+  ['needreverse_1',['needReverse',['../elfFile_8h.html#ab377c00452ebed00f3024890d27ddb7e',1,'utils.c']]],
+  ['newelf32_5fstructure_2',['NewElf32_Structure',['../elfStructure_8h.html#a5e81e8e9f72fd8613ea01fbf2b9e1e8e',1,'elfStructure.c']]],
+  ['newindices_3',['newIndices',['../structElf32__SectionFusion.html#adbc07649d2f6793577b16fcb3889af7d',1,'Elf32_SectionFusion::newIndices()'],['../structElf32__SymbolFusion.html#adbc07649d2f6793577b16fcb3889af7d',1,'Elf32_SymbolFusion::newIndices()'],['../structElf32__RelFusion.html#adbc07649d2f6793577b16fcb3889af7d',1,'Elf32_RelFusion::newIndices()']]],
+  ['newsectionfusion_4',['NewSectionFusion',['../elfStructure_8h.html#a469032e68c22b28f7a91100540b6c9ae',1,'elfStructure.c']]],
+  ['numbersection_5',['numberSection',['../structElf32__SectionFusion.html#ab2db038bfcb955eca04dac929759ce3a',1,'Elf32_SectionFusion']]]
 ];
