+++ doc/html/search/all_b.js
@@ -1,4 +1,16 @@
 var searchData=
 [
-  ['tmpfile_0',['tmpFile',['../structElf32__SectionFusion.html#aa4a41fa3f6cf14b0612fc9817f218ca7',1,'Elf32_SectionFusion']]]
+  ['section_0',['section',['../structElf32__Reim.html#ab217561ce555727ff977aa053656a686',1,'Elf32_Reim']]],
+  ['sectioncount_1',['sectionCount',['../structElf32__Structure.html#a50de16bd16834b393ca054f0c0a74157',1,'Elf32_Structure']]],
+  ['sectiontable_2',['sectionTable',['../structElf32__Structure.html#a83d32eaa69d9281c424ae0f4e48b2224',1,'Elf32_Structure']]],
+  ['showelfheader_3',['ShowElfHeader',['../elfFile_8h.html#a913a20ad3ffeff20afe960474ec7d34b',1,'elfDisplay.c']]],
+  ['showreimplantationtablesanddetails_4',['ShowReimplantationTablesAndDetails',['../elfFile_8h.html#a41b315ed88b2766ba8174206562c419d',1,'elfDisplay.c']]],
+  ['showsectionfromindex_5',['ShowSectionFromIndex',['../elfFile_8h.html#a71973292791ceabf3f391980c1478370',1,'elfDisplay.c']]],
+  ['showsectionfromname_6',['ShowSectionFromName',['../elfFile_8h.html#a2b1d0ae2cfbaef628767f0e3eef27e27',1,'elfDisplay.c']]],
+  ['showsectiontableanddetails_7',['ShowSectionTableAndDetails',['../elfFile_8h.html#a9e335fb66472d849b8246faeb3d51a8c',1,'elfDisplay.c']]],
+  ['showstringfromindex_8',['ShowStringFromIndex',['../utils_8h.html#acd03bf7615ce15f2c30f540e22de08be',1,'utils.c']]],
+  ['showsymbolstableanddetails_9',['ShowSymbolsTableAndDetails',['../elfFile_8h.html#acdb032fddcba7d654a5d0d3456a3a7be',1,'elfDisplay.c']]],
+  ['strtab_10',['strtab',['../structElf32__SymbolFusion.html#aa1f685f471988a9afbecadf56da008c0',1,'Elf32_SymbolFusion']]],
+  ['symbolcount_11',['symbolCount',['../structElf32__Structure.html#afe8b3f4d64938795f407b57b947b5b7a',1,'Elf32_Structure']]],
+  ['symboltable_12',['symbolTable',['../structElf32__Structure.html#a4e261008c5b1db2e7a0a4ce9a2b549cf',1,'Elf32_Structure::symbolTable()'],['../structElf32__SymbolFusion.html#a4e261008c5b1db2e7a0a4ce9a2b549cf',1,'Elf32_SymbolFusion::symbolTable()']]]
 ];
