+++ doc/html/search/functions_1.js
@@ -1,8 +1,4 @@
 var searchData=
 [
-  ['extractelfinformation_0',['ExtractElfInformation',['../elfFile_8h.html#a19ae190d378c6090b327ca358c130861',1,'elfExtract.c']]],
-  ['extractheader_1',['ExtractHeader',['../elfFile_8h.html#a27ba046fdd1470f335e7379cbcab8f50',1,'elfExtract.c']]],
-  ['extractreimplantationtable_2',['ExtractReimplantationTable',['../elfFile_8h.html#a56244c69ae5f37064f10f780b60cb1b7',1,'elfExtract.c']]],
-  ['extractsectiontable_3',['ExtractSectionTable',['../elfFile_8h.html#ac854b60ad3231b4b2618433475e96ee8',1,'elfExtract.c']]],
-  ['extractsymbolstable_4',['ExtractSymbolsTable',['../elfFile_8h.html#a1fffc3393c1db5b07e408be35601b9c0',1,'elfExtract.c']]]
+  ['dofusioncommand_0',['DoFusionCommand',['../elfFile_8h.html#a941f7b8cd416ad6c103dd50a60e3152e',1,'elfFusion.c']]]
 ];
