+++ doc/html/search/functions_3.js
@@ -1,6 +1,9 @@
 var searchData=
 [
-  ['getentrycountfromtype_0',['GetEntryCountFromType',['../utils_8h.html#ae960326a7cdebf3af94ee2ff87076d7c',1,'utils.c']]],
-  ['getsectioncountfromtype_1',['GetSectionCountFromType',['../utils_8h.html#a569d0f2da13d385132706f3f506c37fa',1,'utils.c']]],
-  ['getsectionindexbyname_2',['GetSectionIndexByName',['../utils_8h.html#a2f0ff914ecaef8d1f120ece67cd8678a',1,'utils.c']]]
+  ['freadendian_0',['freadEndian',['../utils_8h.html#a0479d547af675b99a4dd3df59d5d3604',1,'utils.c']]],
+  ['freeelf32_5fstructure_1',['FreeElf32_Structure',['../elfStructure_8h.html#a5a5fb7b2384edf089e3e7c61393e14be',1,'elfStructure.c']]],
+  ['freesectionfusion_2',['FreeSectionFusion',['../elfStructure_8h.html#a747b86ad1ca395db95235b7d51c0c88e',1,'elfStructure.c']]],
+  ['fusionreimplantation_3',['FusionReimplantation',['../elfFile_8h.html#a3144f86bfcb8a1f1b415ba4df07f8688',1,'elfFusion.c']]],
+  ['fusionsections_4',['FusionSections',['../elfFile_8h.html#ab5003ae56f77c5d7fdf49fe349d307a2',1,'elfFusion.c']]],
+  ['fusionsymbols_5',['FusionSymbols',['../elfFile_8h.html#a97d7c187490aea61028153a39adfab6b',1,'elfFile.h']]]
 ];
