+++ doc/html/search/functions_5.js
@@ -1,10 +1,4 @@
 var searchData=
 [
-  ['showelfheader_0',['ShowElfHeader',['../elfFile_8h.html#a913a20ad3ffeff20afe960474ec7d34b',1,'elfDisplay.c']]],
-  ['showreimplantationtablesanddetails_1',['ShowReimplantationTablesAndDetails',['../elfFile_8h.html#a41b315ed88b2766ba8174206562c419d',1,'elfDisplay.c']]],
-  ['showsectionfromindex_2',['ShowSectionFromIndex',['../elfFile_8h.html#a71973292791ceabf3f391980c1478370',1,'elfDisplay.c']]],
-  ['showsectionfromname_3',['ShowSectionFromName',['../elfFile_8h.html#a2b1d0ae2cfbaef628767f0e3eef27e27',1,'elfDisplay.c']]],
-  ['showsectiontableanddetails_4',['ShowSectionTableAndDetails',['../elfFile_8h.html#a9e335fb66472d849b8246faeb3d51a8c',1,'elfDisplay.c']]],
-  ['showstringfromindex_5',['ShowStringFromIndex',['../utils_8h.html#acd03bf7615ce15f2c30f540e22de08be',1,'utils.c']]],
-  ['showsymbolstableanddetails_6',['ShowSymbolsTableAndDetails',['../elfFile_8h.html#acdb032fddcba7d654a5d0d3456a3a7be',1,'elfDisplay.c']]]
+  ['help_0',['help',['../elfFile_8h.html#a97ee70a8770dc30d06c744b24eb2fcfc',1,'elfFile.c']]]
 ];
