+++ doc/html/search/functions_6.js
@@ -1,0 +1,5 @@
+var searchData=
+[
+  ['newelf32_5fstructure_0',['NewElf32_Structure',['../elfStructure_8h.html#a5e81e8e9f72fd8613ea01fbf2b9e1e8e',1,'elfStructure.c']]],
+  ['newsectionfusion_1',['NewSectionFusion',['../elfStructure_8h.html#a469032e68c22b28f7a91100540b6c9ae',1,'elfStructure.c']]]
+];
